+++ spec/memorack_spec.rb
@@ -61,6 +61,7 @@ describe MemoRack do
 					Usage: memorack create [options] PATH
 					       memorack theme  [options] [THEME]
 					       memorack server [options] PATH
+					       memorack build  [options] [PATH]
 
 					    -h, --help                       Show this message
 				EOD
@@ -92,6 +93,19 @@ describe MemoRack do
 					    -h, --help                       Show this message
 				EOD
 			end
+
+			it "build" do
+				proc { memorack 'build', '-h' }.must_output nil, <<-EOD.cut_indent
+					Usage: memorack build [options] [PATH]
+
+					    -o, --output DIRECTORY           Output directory (default: _site)
+					    -t, --theme THEME                use THEME (default: custom)
+					        --url URL                    Site URL (default: )
+					        --local                      Site URL is output directory
+					        --prettify                   prettify URL
+					    -h, --help                       Show this message
+				EOD
+			end
 		end
 
 		describe "ja" do
@@ -104,6 +118,7 @@ describe MemoRack do
 					Usage: memorack create [options] PATH
 					       memorack theme  [options] [THEME]
 					       memorack server [options] PATH
+					       memorack build  [options] [PATH]
 
 					    -h, --help                       このメッセージを表示
 				EOD
@@ -133,6 +148,19 @@ describe MemoRack do
 
 					    -p, --port PORT                  ポートを使う (省略値: 9292)
 					    -t, --theme THEME                テーマを使う (省略値: oreilly)
+					    -h, --help                       このメッセージを表示
+				EOD
+			end
+
+			it "build" do
+				proc { memorack 'build', '-h' }.must_output nil, <<-EOD.cut_indent
+					Usage: memorack build [options] [PATH]
+
+					    -o, --output DIRECTORY           出力するディレクトリ (省略値: _site)
+					    -t, --theme THEME                テーマを使う (省略値: custom)
+					        --url URL                    サイトURL (省略値: )
+					        --local                      サイトURLをアウトプットディレクトリにする
+					        --prettify                   綺麗なURLになるように生成する
 					    -h, --help                       このメッセージを表示
 				EOD
 			end
@@ -210,13 +238,14 @@ describe MemoRack do
 
 				`cd themes/#{theme}; find . -print`.must_equal <<-EOD.cut_indent
 					.
-					./2-column.scss
 					./404.md
-					./basic-styles.scss
 					./config.json
+					./css
+					./css/2-column.scss
+					./css/basic-styles.scss
+					./css/styles.scss
 					./error.html
 					./index.html
-					./styles.scss
 				EOD
 			}
 		end
@@ -229,10 +258,179 @@ describe MemoRack do
 				proc { memorack 'theme', '-c', File.join(theme, fname) }.must_output "Created '#{fname}'\n"
 			}
 		end
+
+	end
+
+	describe "build" do
+		before do
+			@hash = {}
+			@hash['basic']		= 'd414b3942f61e6b0a07f6458bfb133da40e7c7b8'
+			@hash['oreilly']	= '56b0c0f6a7394904442d7af6986797b1300301c3'
+
+			@file_lists = <<-EOD.cut_indent
+				.
+				./css
+				./css/2-column.css
+				./css/basic-styles.css
+				./css/styles.css
+				./index.html
+				./README.html
+			EOD
+		end
+
+		it "build" do
+			theme  = 'oreilly'
+			output = '_site'
+
+			chmemo { |name|
+				proc { memorack 'build' }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+				}
+			}
+		end
+
+		it "build PATH" do
+			theme  = 'basic'
+			output = '_site'
+
+			chmemo { |name|
+				dirname = 'data'
+				File.rename('content', dirname)
+
+				Dir.chdir('..')
+
+				path = File.join(name, dirname)
+				proc { memorack 'build', path }.must_output "Build '#{path}' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+				}
+			}
+		end
+
+		it "build --output DIRECTORY" do
+			theme  = 'oreilly'
+			output = 'output'
+
+			chmemo { |name|
+				proc { memorack 'build', '--output', output }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+				}
+			}
+		end
+
+		it "build --theme THEME" do
+			theme  = 'basic'
+			output = '_site'
+
+			chmemo { |name|
+				proc { memorack 'build', '--theme', theme }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+				}
+			}
+		end
+
+		it "build --url URL" do
+			theme  = 'oreilly'
+			output = '_site'
+			url    = 'http://memo.pow'
+
+			chmemo { |name|
+				proc { memorack 'build', '--url', url }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+					File.read('index.html').must_match %r[<a href="#{url}/README.html">README</a>]
+				}
+			}
+		end
+
+		it "build --local" do
+			theme  = 'oreilly'
+			output = '_site'
+
+			chmemo { |name|
+				url = 'file://' + File.expand_path(output)
+
+				proc { memorack 'build', '--local' }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal @file_lists
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+					File.read('index.html').must_match %r[<a href="#{url}/README.html">README</a>]
+				}
+			}
+		end
+
+		it "build --prettify" do
+			theme  = 'oreilly'
+			output = '_site'
+			url    = ''
+
+			chmemo { |name|
+				proc { memorack 'build', '--prettify' }.must_output "Build 'content/' -> '#{output}'\n"
+
+				Dir.chdir(output) { |output|
+					`find . -print`.must_equal <<-EOD.cut_indent
+						.
+						./css
+						./css/2-column.css
+						./css/basic-styles.css
+						./css/styles.css
+						./index.html
+						./README
+						./README/index.html
+					EOD
+
+					`git hash-object css/styles.css`.must_equal @hash[theme]+"\n"
+					File.read('index.html').must_match %r[<a href="#{url}/README">README</a>]
+				}
+			}
+		end
 	end
 
 	describe "server" do
-		it "server"
+		include Rack::Test::Methods
+
+		before do
+			@cwd = Dir.pwd
+
+			name = 'memo'
+			Dir.chdir(@tmpdir)
+			proc { memorack 'create', name }.must_output "Created '#{name}'\n"
+			Dir.chdir(name)
+		end
+
+		def app
+			Rack::Builder.parse_file('config.ru').first
+		end
+
+		it "server /" do
+			get '/'
+			last_response.must_be :ok?
+			last_response.body.must_match /Powered by <a href="#{MemoRack::HOMEPAGE}"[^>]*>/
+		end
+
+		it "server /README" do
+			get '/README'
+			last_response.must_be :ok?
+			last_response.body.must_match /このディレクトリに markdown 等のメモファイルを作成してください/
+		end
+
+		after do
+			Dir.chdir(@cwd)
+		end
 	end
 
 	after do