+++ HISTORY.md
@@ -1,5 +1,11 @@
 ## History
 
+### v0.1.0 / 2013-04-15
+
+* generate HTMLs for static site
+* add config.json options(requires, public, site.url)
+* change basic theme directory structure
+
 ### v0.0.4 / 2013-04-12
 
 * fix link in content